Set resolve to be a task, triggerable by a button
