+++ plenario/templates/apiary/index.html
@@ -6,23 +6,24 @@
         {% for element in elements %}
             <li class="list-group-item">
                 <div class="row">
-                    <div class="col-md-12">
+                    <div class="col-md-6">
                         <h4>{{ element.datetime }}</h4>
+                    </div>
+                    <div class="col-md-6">
+                        <h4>
+                            <a href="apiary/resolve/{{ element.sensor }}" type="button" class="btn btn-primary pull-right">
+                                Resolve {{ element.sensor }}
+                            </a>
+                        </h4>
                     </div>
                 </div>
                 <div class="row">
                     <div class="col-md-6">
                         <h5>{{ element.sensor }}</h5>
-                    </div>
-                    <div class="col-md-6">
-                        <h5>Incoming</h5>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-md-6">
                         <pre>{{ element.expected }}</pre>
                     </div>
                     <div class="col-md-6">
+                        <h5>Incoming</h5>
                         <pre>{{ element.incoming }}</pre>
                     </div>
                 </div>
